Finished preliminary signature verification
Public key is downloaded to /etc/butterknife/trusted.gpg.d
from originating server if necessary. Manifests are downloaded to
/var/lib/butterknife/manifests/@template:xxx:yyy:zzz if necessary.
Similarly GPG signature is downloaded from originating server to
/var/lib/butterknife/manifests/@template:xxx:yyy:zzz.asc
if necessary. Temporary files are used to guarantee atomicity
of write operations. Manifest of the subvolume is generated
on the fly and compared to signed manifest, finally
gpgv is used to check that manifest matches the signature.